#1 - removed most of the logic from Main.scala
